+++ CHANGELOG.md
@@ -21,6 +21,9 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 - updates `orbit help` command to have latest manuals and available subcommands to display manual pages
 - adds `--no-clean` flag to `orbit build` and `orbit test` to prevent automatic cleaning before build process
 
+### Fixes
+- resolves bug where string swapping would not occur on source url if using a custom protocol
+
 ### Docs
 - adds architecture diagram for topics overview page
 - updates command workflow diagram to use newer commands